+++ src/layout/navBars/breadcrumb/breadcrumb.vue
@@ -7,10 +7,12 @@
 			@click="onThemeConfigChange"
 		/>
 		<el-breadcrumb class="layout-navbars-breadcrumb-hide">
-			<transition-group name="breadcrumb" mode="out-in">
-				<el-breadcrumb-item v-for="(v, k) in breadcrumbList" :key="v.meta.title">
+			<transition-group name="breadcrumb">
+				<el-breadcrumb-item v-for="(v, k) in breadcrumbList" :key="!v.meta.tagsViewName ? v.meta.title : v.meta.tagsViewName">
 					<span v-if="k === breadcrumbList.length - 1" class="layout-navbars-breadcrumb-span">
-						<SvgIcon :name="v.meta.icon" class="layout-navbars-breadcrumb-iconfont" v-if="themeConfig.isBreadcrumbIcon" />{{ $t(v.meta.title) }}
+						<SvgIcon :name="v.meta.icon" class="layout-navbars-breadcrumb-iconfont" v-if="themeConfig.isBreadcrumbIcon" />
+						<div v-if="!v.meta.tagsViewName">{{ $t(v.meta.title) }}</div>
+						<div v-else>{{ v.meta.tagsViewName }}</div>
 					</span>
 					<a v-else @click.prevent="onBreadcrumbClick(v)">
 						<SvgIcon :name="v.meta.icon" class="layout-navbars-breadcrumb-iconfont" v-if="themeConfig.isBreadcrumbIcon" />{{ $t(v.meta.title) }}
@@ -25,6 +27,7 @@
 import { toRefs, reactive, computed, onMounted, defineComponent } from 'vue';
 import { onBeforeRouteUpdate, useRoute, useRouter } from 'vue-router';
 import { Local } from '/@/utils/storage';
+import other from '/@/utils/other';
 import { storeToRefs } from 'pinia';
 import { useThemeConfig } from '/@/stores/themeConfig';
 import { useRoutesList } from '/@/stores/routesList';
@@ -77,8 +80,8 @@ export default defineComponent({
 		};
 		// 处理面包屑数据
 		const getBreadcrumbList = (arr: Array<string>) => {
-			arr.map((item: any) => {
-				state.routeSplit.map((v: any, k: number, arrs: any) => {
+			arr.forEach((item: any) => {
+				state.routeSplit.forEach((v: any, k: number, arrs: any) => {
 					if (state.routeSplitFirst === item.path) {
 						state.routeSplitFirst += `/${arrs[state.routeSplitIndex]}`;
 						state.breadcrumbList.push(item);
@@ -90,6 +93,7 @@ export default defineComponent({
 		};
 		// 当前路由字符串切割成数组，并删除第一项空内容
 		const initRouteSplit = (path: string) => {
+			state.breadcrumbList = [];
 			if (!themeConfig.value.isBreadcrumb) return false;
 			state.breadcrumbList = [routesList.value[0]];
 			state.routeSplit = path.split('/');
@@ -97,6 +101,8 @@ export default defineComponent({
 			state.routeSplitFirst = `/${state.routeSplit[0]}`;
 			state.routeSplitIndex = 1;
 			getBreadcrumbList(routesList.value);
+			if (route.name === 'home') state.breadcrumbList.shift();
+			state.breadcrumbList[state.breadcrumbList.length - 1].meta.tagsViewName = other.setTagsViewNameI18n(route);
 		};
 		// 页面加载时
 		onMounted(() => {